backend/requirements.txt
# Python packages
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
opencv-python-headless==4.5.3.56
face-recognition==1.3.0
numpy>=1.24.0
python-dotenv==0.19.0
Pillow==8.3.2
bcrypt==3.2.0
